refactor: Notification 도메인을 app.bottlenote 패키지로 이전
