+++ bottlenote-mono/src/main/java/app/bottlenote/notification/exception/NotificationException.java
@@ -1,0 +1,10 @@
+package app.bottlenote.notification.exception;
+
+import app.bottlenote.global.exception.custom.AbstractCustomException;
+
+public class NotificationException extends AbstractCustomException {
+
+  public NotificationException(NotificationExceptionCode code) {
+    super(code);
+  }
+}
