--- bottlenote-mono/src/main/java/app/external/notification/exception/NotificationException.java
@@ -1,10 +1,0 @@
-package app.external.notification.exception;
-
-import app.bottlenote.global.exception.custom.AbstractCustomException;
-
-public class NotificationException extends AbstractCustomException {
-
-  public NotificationException(NotificationExceptionCode code) {
-    super(code);
-  }
-}
