+++ bottlenote-mono/src/test/java/app/bottlenote/notification/application/UserNotificationServiceTest.java
@@ -1,0 +1,220 @@
+package app.bottlenote.notification.application;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+import app.bottlenote.notification.data.payload.NotificationMessage;
+import app.bottlenote.notification.domain.Notification;
+import app.bottlenote.notification.domain.constant.NotificationCategory;
+import app.bottlenote.notification.domain.constant.NotificationStatus;
+import app.bottlenote.notification.domain.constant.NotificationType;
+import app.bottlenote.notification.exception.NotificationException;
+import app.bottlenote.notification.exception.NotificationExceptionCode;
+import app.bottlenote.notification.fixture.InMemoryNotificationRepository;
+import app.bottlenote.user.constant.GenderType;
+import app.bottlenote.user.constant.SocialType;
+import app.bottlenote.user.constant.UserType;
+import app.bottlenote.user.domain.User;
+import app.bottlenote.user.exception.UserException;
+import app.bottlenote.user.exception.UserExceptionCode;
+import app.bottlenote.user.fixture.InMemoryUserQueryRepository;
+import java.util.List;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.Test;
+
+@Tag("unit")
+@DisplayName("UserNotificationService 단위 테스트")
+class UserNotificationServiceTest {
+
+  private static final Long USER_ID = 1L;
+  private static final Long OTHER_USER_ID = 2L;
+
+  private InMemoryUserQueryRepository userRepository;
+  private InMemoryNotificationRepository notificationRepository;
+  private UserNotificationService service;
+
+  @BeforeEach
+  void setUp() {
+    userRepository = new InMemoryUserQueryRepository();
+    notificationRepository = new InMemoryNotificationRepository();
+    service = new UserNotificationService(userRepository, notificationRepository);
+  }
+
+  @Nested
+  @DisplayName("알림을 생성할 때")
+  class SendNotification {
+
+    @Test
+    @DisplayName("대상 사용자가 있으면 알림을 저장한다")
+    void sendNotification_whenUserExists_savesNotification() {
+      seedUser(USER_ID);
+
+      service.sendNotification(
+          NotificationMessage.create(
+              USER_ID,
+              NotificationType.USER,
+              NotificationCategory.REVIEW,
+              "새 댓글",
+              "리뷰에 댓글이 달렸습니다."));
+
+      assertThat(notificationRepository.findAll())
+          .hasSize(1)
+          .first()
+          .satisfies(
+              notification -> {
+                assertThat(notification.getUserId()).isEqualTo(USER_ID);
+                assertThat(notification.getTitle()).isEqualTo("새 댓글");
+                assertThat(notification.getContent()).isEqualTo("리뷰에 댓글이 달렸습니다.");
+                assertThat(notification.getType()).isEqualTo(NotificationType.USER);
+                assertThat(notification.getCategory()).isEqualTo(NotificationCategory.REVIEW);
+                assertThat(notification.getIsRead()).isFalse();
+              });
+    }
+
+    @Test
+    @DisplayName("대상 사용자가 없으면 예외를 던진다")
+    void sendNotification_whenUserMissing_throwsException() {
+      NotificationMessage message =
+          NotificationMessage.create(
+              USER_ID, NotificationType.USER, NotificationCategory.REVIEW, "제목", "내용");
+
+      assertThatThrownBy(() -> service.sendNotification(message))
+          .isInstanceOf(UserException.class)
+          .extracting(ex -> ((UserException) ex).getExceptionCode())
+          .isEqualTo(UserExceptionCode.NOTIFICATION_USER_NOT_FOUND);
+    }
+  }
+
+  @Nested
+  @DisplayName("알림함을 조회할 때")
+  class InboxQuery {
+
+    @Test
+    @DisplayName("사용자별 알림을 최신 id 순으로 반환한다")
+    void getNotifications_whenMultipleExist_returnsOwnOrderedByIdDesc() {
+      Notification older = seedNotification(USER_ID, "old");
+      Notification newer = seedNotification(USER_ID, "new");
+      seedNotification(OTHER_USER_ID, "other");
+
+      List<Notification> result = service.getNotifications(USER_ID);
+
+      assertThat(result)
+          .extracting(Notification::getId)
+          .containsExactly(newer.getId(), older.getId());
+      assertThat(result).extracting(Notification::getTitle).containsExactly("new", "old");
+    }
+
+    @Test
+    @DisplayName("미읽음 개수를 사용자 기준으로 센다")
+    void countUnread_whenMixedReadState_countsOnlyUnreadOfUser() {
+      Notification unread = seedNotification(USER_ID, "unread-1");
+      Notification read = seedNotification(USER_ID, "read");
+      service.markAsRead(USER_ID, read.getId());
+      seedNotification(USER_ID, "unread-2");
+      seedNotification(OTHER_USER_ID, "other-unread");
+
+      long count = service.countUnread(USER_ID);
+
+      assertThat(count).isEqualTo(2L);
+      assertThat(unread.getIsRead()).isFalse();
+    }
+  }
+
+  @Nested
+  @DisplayName("알림을 읽음 처리할 때")
+  class MarkRead {
+
+    @Test
+    @DisplayName("본인 알림 단건을 읽음 처리한다")
+    void markAsRead_whenOwnNotification_marksRead() {
+      Notification notification = seedNotification(USER_ID, "title");
+
+      service.markAsRead(USER_ID, notification.getId());
+
+      Notification saved = notificationRepository.findById(notification.getId()).orElseThrow();
+      assertThat(saved.getIsRead()).isTrue();
+      assertThat(saved.getStatus()).isEqualTo(NotificationStatus.READ);
+    }
+
+    @Test
+    @DisplayName("타 사용자 알림은 찾을 수 없음으로 처리한다")
+    void markAsRead_whenOtherUsersNotification_throwsNotFound() {
+      Notification other = seedNotification(OTHER_USER_ID, "other");
+
+      assertThatThrownBy(() -> service.markAsRead(USER_ID, other.getId()))
+          .isInstanceOf(NotificationException.class)
+          .extracting(ex -> ((NotificationException) ex).getExceptionCode())
+          .isEqualTo(NotificationExceptionCode.NOTIFICATION_NOT_FOUND);
+      assertThat(other.getIsRead()).isFalse();
+    }
+
+    @Test
+    @DisplayName("존재하지 않는 알림이면 예외를 던진다")
+    void markAsRead_whenMissing_throwsNotFound() {
+      assertThatThrownBy(() -> service.markAsRead(USER_ID, 999L))
+          .isInstanceOf(NotificationException.class)
+          .extracting(ex -> ((NotificationException) ex).getExceptionCode())
+          .isEqualTo(NotificationExceptionCode.NOTIFICATION_NOT_FOUND);
+    }
+
+    @Test
+    @DisplayName("본인 미읽음 알림을 모두 읽음 처리한다")
+    void markAllAsRead_whenUnreadExist_marksAllOwn() {
+      Notification first = seedNotification(USER_ID, "a");
+      Notification second = seedNotification(USER_ID, "b");
+      Notification other = seedNotification(OTHER_USER_ID, "other");
+
+      int updated = service.markAllAsRead(USER_ID);
+
+      assertThat(updated).isEqualTo(2);
+      assertThat(first.getIsRead()).isTrue();
+      assertThat(second.getIsRead()).isTrue();
+      assertThat(other.getIsRead()).isFalse();
+      assertThat(service.countUnread(USER_ID)).isZero();
+      assertThat(service.countUnread(OTHER_USER_ID)).isEqualTo(1L);
+    }
+
+    @Test
+    @DisplayName("이미 모두 읽은 경우 0건을 반환한다")
+    void markAllAsRead_whenNoneUnread_returnsZero() {
+      Notification notification = seedNotification(USER_ID, "a");
+      service.markAsRead(USER_ID, notification.getId());
+
+      int updated = service.markAllAsRead(USER_ID);
+
+      assertThat(updated).isZero();
+    }
+  }
+
+  private void seedUser(Long userId) {
+    User user =
+        User.builder()
+            .email("user" + userId + "@example.com")
+            .nickName("user" + userId)
+            .age(25)
+            .gender(GenderType.MALE)
+            .socialType(List.of(SocialType.KAKAO))
+            .role(UserType.ROLE_USER)
+            .build();
+    userRepository.save(user);
+    // InMemoryUserQueryRepository assigns sequential ids starting at 1.
+    if (!user.getId().equals(userId)) {
+      throw new IllegalStateException(
+          "테스트 픽스처 userId 불일치: expected=" + userId + ", actual=" + user.getId());
+    }
+  }
+
+  private Notification seedNotification(Long userId, String title) {
+    return notificationRepository.save(
+        Notification.builder()
+            .userId(userId)
+            .title(title)
+            .content(title + "-content")
+            .type(NotificationType.USER)
+            .category(NotificationCategory.REVIEW)
+            .build());
+  }
+}
